requirements.txt
requests>=2.28.0
Pillow>=10.0.0
tqdm>=4.64.0
PyMuPDF>=1.23.0 